"Prepared" for confirmation mails
